requirements.txt
requests
whois
colorama
scikit-image
virustotal-python
selenium
Pillow
